clean up some stuff, add some c++ datastructures, start some work

+++ utils/tools/update-diskimage.sh
@@ -51,15 +51,19 @@ pushd $PROJECT_DIR/build > /dev/null
 		exit 1
 	fi
 
-	mcopy -D o -snQ -i esp-part.img $PROJECT_DIR/utils/refind.conf ::/EFI/BOOT/refind.conf
+	if [ $PROJECT_DIR/utils/refind.conf -nt esp-part.img ]; then
+		dd if=esp-part.img of=disk.img bs=$SECTOR_SIZE count=$(expr $ESP_PARTITION_END - $ESP_PARTITION_START + 1) seek=$ESP_PARTITION_START conv=notrunc status=none
 
-	mdeltree -i root-part.img ::/*
-	mcopy -D o -snQ -i root-part.img $PROJECT_DIR/build/sysroot/* ::/
+	fi
 
-	# join the two partitions together.
-	dd if=esp-part.img of=disk.img bs=$SECTOR_SIZE count=$(expr $ESP_PARTITION_END - $ESP_PARTITION_START + 1) seek=$ESP_PARTITION_START conv=notrunc status=none
-	dd if=root-part.img of=disk.img bs=$SECTOR_SIZE count=$(expr $ROOT_PARTITION_END - $ROOT_PARTITION_START + 1) seek=$ROOT_PARTITION_START conv=notrunc status=none
+	ESP_IMG_OFFSET=disk.img@@$(expr $ESP_PARTITION_START \* $SECTOR_SIZE)
+	ROOT_IMG_OFFSET=disk.img@@$(expr $ROOT_PARTITION_START \* $SECTOR_SIZE)
 
+
+	mcopy -D o -snQ -i $ESP_IMG_OFFSET $PROJECT_DIR/utils/refind.conf ::/EFI/BOOT/refind.conf
+
+	mdeltree -i $ROOT_IMG_OFFSET ::/*
+	mcopy -D o -snQ -i $ROOT_IMG_OFFSET $PROJECT_DIR/build/sysroot/* ::/
 
 popd > /dev/null
